requirements.txt
python-telegram-bot==20.7
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.0
google-generativeai==0.3.2
protobuf==4.25.3